CHROMIUM: wpa_supplicant: Clear blacklist when SSID configs change
If the stored configurations for an SSID have changed, we can no longer
trust the current blacklist state of that SSID, since the updated
configs could change the behavior of the network. E.g. the station could
be blacklisted due to a bad password, and the config could be updated to
store the correct password. In this case, keeping the station in the
blacklist will prevent the user from connecting to the station after the
correct password has been updated.

Add the value blacklist_out_of_date to the wpa_ssid struct. Update this
value every time a config is changed through wpa_set_config. Check this
value in wpa_blacklist_get() to clear the blacklist whenever the
configs of current_ssid have changed.

This solution was chosen over simply clearing the blacklist whenever
configs change because the user should be able to change configs on an
inactive SSID without affecting the blacklist for the currently active
SSID. This way, the blacklist won't be cleared until the user attempts
to connect to the inactive network again. Furthermore, the blacklist is
stored per-BSSID while configs are stored per-SSID, so we don't have the
option to just clear out certain blacklist entries that would be
affected by the configs.

Finally, the function wpa_supplicant_reload_configuration() causes the
configs to be reloaded from scratch, so after a call to this function
all bets are off as to the relevance of our current blacklist state.
Thus, we clear the entire blacklist within this function.

Signed-off-by: Kevin Lund <kglund@google.com>
Signed-off-by: Brian Norris <briannorris@chromium.org>
(am from https://patchwork.ozlabs.org/patch/1307849)

Conflict was from added SSID config parameters in the upstream tree.

BUG=chromium:1040974, chromium:1051374
TEST=network_WiFi_SimpleConnect.wifi_checkWPA2
Refactor test to give dut bad password, blacklist is cleared correctly
when the password is updated, and a connection is made as expected.
TEST=Create simple config file at /etc/wpa_supplicant/wpa_supplicant.conf
(dut) wpa_supplicant -B -i wlan0 -c /etc/wpa_supplicant/wpa_supplicant.conf
(dut) wpa_cli
(wpa_cli) blacklist ${BSSID}
(wpa_cli) reconfigure
(wpa_cli) blacklist
Check that reconfigure caused the blacklist to clear.

Change-Id: I07433a2c49b6674fd0e96ea825bf55beae510742
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2242374
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>