FROMLIST: global parser functions return 1 when property unchanged
Currently, wpa_config_set, the function that sets per-network
properties, returns 1 when a property it attempts to set is unchanged.
Its global parallel, wpa_config_process_global, doesn't do this even
though much of the code is very similar. Change this, and several of the
parser functions, to resemble the per-network parser and setter
functions.

BUG=chromium:1059866
TEST=use gdbus and wpa_cli to set properties to the same value they're
already set to, check to make sure that 1 was returned.

(am from http://patchwork.ozlabs.org/patch/1318163)

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: Iad08bdc28291b916e2f1dc085c548668e25398bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2284181
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>