FROMLIST: skip property update actions when wpa_config_set returns 1
When network properties are updated via dbus, wpa_config_set is used to
update the property in the wpa_ssid struct. If it returns 1, the
property was not changed and there's no need to perform any of the
update actions.

BUG=chromium:1059866
TEST=set psk of the associated network to the same value via gdbus and
check logs to make sure nothing happens.

(am from http://patchwork.ozlabs.org/patch/1318162)

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: Id73e3e738d5f98cb3315d3272831e383880ec50d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2271000
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>